refactor: Adds 'theme' folder for missing theme styles - #23
